Move compute_M and sane_parameters to utils.py
