+++ tests/test_fluctuation_dissipation.py
@@ -6,31 +6,7 @@ from numpy.linalg import eig
 import logging
 
 from libMobility import *
-
-sane_parameters = {
-    "PSE": {"psi": 1.0, "Lx": 32, "Ly": 32, "Lz": 32, "shearStrain": 0.0},
-    "NBody": {"algorithm": "advise"},
-    "DPStokes": {"dt": 1, "Lx": 16, "Ly": 16, "zmin": -6, "zmax": 6},
-    "SelfMobility": {"parameter": 5.0},
-}
-
-
-def compute_M(solver, numberParticles):
-
-    precision = np.float32 if solver.precision == "float" else np.float64
-
-    size = 3 * numberParticles
-    M = np.zeros((size, size), dtype=precision)
-
-    I = np.identity(size, dtype=precision)
-    for i in range(0, size):
-        forces = I[:, i].copy()
-        mf = np.zeros(size, dtype=precision)
-        solver.Mdot(forces, mf)
-
-        M[:, i] = mf
-
-    return M
+from utils import compute_M, sane_parameters
 
 
 def fluctuation_dissipation_KS(M, fluctuation_method):
@@ -75,8 +51,11 @@ def fluctuation_dissipation_KS(M, fluctuation_method):
         (DPStokes, ("periodic", "periodic", "two_walls")),
     ],
 )
-@pytest.mark.parametrize("hydrodynamicRadius", [1.0, 0.95, 1.12])
-def test_fluctuation_dissipation(Solver, periodicity, hydrodynamicRadius):
+@pytest.mark.parametrize("hydrodynamicRadius", [0.95, 1.12])
+@pytest.mark.parametrize("numberParticles", [1, 2, 3, 10])
+def test_fluctuation_dissipation(
+    Solver, periodicity, hydrodynamicRadius, numberParticles
+):
 
     precision = np.float32 if Solver.precision == "float" else np.float64
